+++ src/core/index.ts
@@ -119,10 +119,7 @@ export {
   type MorphLifecycleOptions,
 } from "./morph-lifecycle.js"
 export * from "./parser.js"
-export {
-  type ExpoTurboModuleVersion,
-  serializeModuleVersionsHeader,
-} from "./protocol-request.js"
+export { serializeClientDescriptor } from "./protocol-request.js"
 export * from "./recent-request-ids.js"
 export {
   BeforeFetchRequestEvent,